fix: diagnostic report table, not obs table
